wip: Migrate tests from extrafields with errors.
