+++ ckanext/ontario_theme/tests/test_helpers.py
@@ -6,7 +6,10 @@ import nose.tools
 import json
 import os
 from ckan.common import config
-from ckanext.ontario_theme.plugin import (get_license)
+from ckanext.ontario_theme.plugin import (
+    get_license,
+    default_locale
+)
 assert_equals = nose.tools.assert_equals
 
 
@@ -19,3 +22,9 @@ class TestGetLicense(object):
         ) as licenses:
             license = json.load(licenses)[0]
         assert_equals(get_license("OGL-ON-1.0")._data, license)
+
+
+class TestDefaultLocale(object):
+    def test_default_locale_returns_proper_value(self):
+        default = config.get('ckan.locale_default', 'en')
+        assert_equals(default_locale(), default)
